# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: متجر الأفق المتجدد API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/summary:
    get:
      operationId: adminSummary
      tags: [admin]
      summary: Dashboard summary
      parameters:
        - $ref: "#/components/parameters/SessionToken"
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminSummary" }
  /admin/orders:
    get:
      operationId: adminListOrders
      tags: [admin]
      summary: List all orders
      parameters:
        - $ref: "#/components/parameters/SessionToken"
        - in: query
          name: status
          schema: { $ref: "#/components/schemas/OrderStatus" }
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Order" }
  /admin/orders/{id}/status:
    patch:
      operationId: adminUpdateOrderStatus
      tags: [admin]
      summary: Update order status
      parameters:
        - $ref: "#/components/parameters/SessionToken"
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateOrderStatusBody" }
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
  /admin/orders/{id}/stockout:
    post:
      operationId: adminFlagStockout
      tags: [admin]
      summary: Flag an order item as out of stock
      parameters:
        - $ref: "#/components/parameters/SessionToken"
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FlagStockoutBody" }
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List users
      parameters:
        - $ref: "#/components/parameters/SessionToken"
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/User" }
  /admin/users/{id}/role:
    patch:
      operationId: adminUpdateUserRole
      tags: [admin]
      summary: Promote or demote a user
      parameters:
        - $ref: "#/components/parameters/SessionToken"
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateUserRoleBody" }
      responses:
        "200":
          description: User
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
  /admin/products:
    get:
      operationId: adminListProducts
      tags: [admin]
      summary: List all products
      parameters:
        - $ref: "#/components/parameters/SessionToken"
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Product" }
    post:
      operationId: adminCreateProduct
      tags: [admin]
      summary: Create a product
      parameters:
        - $ref: "#/components/parameters/SessionToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateProductBody" }
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
...
  /admin/categories:
    get:
      operationId: adminListCategories
      tags: [admin]
      summary: List all categories (admin)
      parameters:
        - $ref: "#/components/parameters/SessionToken"
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Category" }
    post:
      operationId: adminCreateCategory
      tags: [admin]
      summary: Create a category
      parameters:
        - $ref: "#/components/parameters/SessionToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminCreateCategoryBody" }
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
  /admin/categories/{id}:
    put:
      operationId: adminUpdateCategory
      tags: [admin]
      summary: Update a category
      parameters:
        - $ref: "#/components/parameters/SessionToken"
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminUpdateCategoryBody" }
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
    delete:
      operationId: adminDeleteCategory
      tags: [admin]
      summary: Delete a category
      parameters:
        - $ref: "#/components/parameters/SessionToken"
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
...
  /admin/products/{productId}/colors:
    get:
      operationId: adminGetProductColors
      tags: [admin]
      summary: Get all colors for a product (admin)
      parameters:
        - $ref: "#/components/parameters/SessionToken"
        - in: path
          name: productId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Colors
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ProductColor" }
    post:
      operationId: adminCreateProductColor
      tags: [admin]
      summary: Add a color variant to a product
      parameters:
        - $ref: "#/components/parameters/SessionToken"
        - in: path
          name: productId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateProductColorBody" }
      responses:
        "200":
          description: Color
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductColor" }
  /admin/product-colors/{colorId}:
    patch:
      operationId: adminUpdateProductColor
      tags: [admin]
      summary: Update a product color
      parameters:
        - $ref: "#/components/parameters/SessionToken"
        - in: path
          name: colorId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateProductColorBody" }
      responses:
        "200":
          description: Color
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductColor" }
    delete:
      operationId: adminDeleteProductColor
      tags: [admin]
      summary: Delete a product color
      parameters:
        - $ref: "#/components/parameters/SessionToken"
        - in: path
          name: colorId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
...
  /admin/about:
    put:
      operationId: adminUpdateAboutPage
      tags: [admin]
      summary: Update about page content
      parameters:
        - $ref: "#/components/parameters/SessionToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AboutPageContent" }
      responses:
        "200":
          description: Updated content
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AboutPageContent" }
  /admin/products/{id}:
    patch:
      operationId: adminUpdateProduct
      tags: [admin]
      summary: Update a product
      parameters:
        - $ref: "#/components/parameters/SessionToken"
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateProductBody" }
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
    delete:
      operationId: adminDeleteProduct
      tags: [admin]
      summary: Delete a product
      parameters:
        - $ref: "#/components/parameters/SessionToken"
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
components:
  parameters:
    SessionToken:
      in: header
      name: x-session-token
      required: false
      schema: { type: string }
  schemas:
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
...
    Gender:
      type: string
      enum: [male, female]
    UserRole:
      type: string
      enum: [customer, admin]
...
    User:
      type: object
      required: [id, fullName, phone, region, age, gender, role]
      properties:
        id: { type: integer }
        fullName: { type: string }
        phone: { type: string }
        region: { type: string }
        age: { type: integer }
        gender: { $ref: "#/components/schemas/Gender" }
        role: { $ref: "#/components/schemas/UserRole" }
        avatarUrl: { type: string, nullable: true }
        walletBalance: { type: number }
        createdAt: { type: string }
...
    Category:
      type: object
      required: [id, slug, name]
      properties:
        id: { type: integer }
        slug: { type: string }
        name: { type: string }
        nameEn: { type: string, nullable: true }
        icon: { type: string, nullable: true }
        imageUrl: { type: string, nullable: true }
...
    ProductColor:
      type: object
      required: [id, productId, name, hexCode, stock, enabled, sortOrder]
      properties:
        id: { type: integer }
        productId: { type: integer }
        name: { type: string }
        hexCode: { type: string }
        stock: { type: integer }
        enabled: { type: boolean }
        sortOrder: { type: integer }
    Product:
      type: object
      required: [id, name, price, description, imageUrl, stock, subcategoryId, subcategorySlug, subcategoryName, categorySlug, categoryName]
      properties:
        id: { type: integer }
        name: { type: string }
        price: { type: number }
        description: { type: string }
        imageUrl: { type: string }
        image2: { type: string, nullable: true }
        image3: { type: string, nullable: true }
        colorsEnabled: { type: boolean }
        stock: { type: integer }
        subcategoryId: { type: integer }
        subcategorySlug: { type: string }
        subcategoryName: { type: string }
        categorySlug: { type: string }
        categoryName: { type: string }
        createdAt: { type: string }
        averageRating: { type: number, nullable: true }
        ratingCount: { type: integer }
        colors:
          type: array
          items: { $ref: "#/components/schemas/ProductColor" }
...
    OrderStatus:
      type: string
      description: "حالة الطلب"
      enum: [pending, preparing, prepared, delivered_to_courier, cancelled, stockout, payment_received]
...
    CreateProductColorBody:
      type: object
      required: [name, hexCode, stock]
      properties:
        name: { type: string }
        hexCode: { type: string }
        stock: { type: integer, minimum: 0 }
        enabled: { type: boolean }
        sortOrder: { type: integer }
    UpdateProductColorBody:
      type: object
      properties:
        name: { type: string }
        hexCode: { type: string }
        stock: { type: integer, minimum: 0 }
        enabled: { type: boolean }
        sortOrder: { type: integer }
...
    AboutPageContent:
      type: object
      properties:
        managerName: { type: string }
        managerTitle: { type: string }
        managerBio: { type: string }
        managerAvatarUrl: { type: string }
        companyDescription: { type: string }
        facebookUrl: { type: string }
        instagramUrl: { type: string }
        tiktokUrl: { type: string }
        branches:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              location: { type: string }
              description: { type: string }
              imageUrl: { type: string }
              managerName: { type: string }
              managerTitle: { type: string }
    OrderItem:
      type: object
      required: [id, productId, productName, productImage, price, quantity, outOfStock]
      properties:
        id: { type: integer }
        productId: { type: integer }
        productName: { type: string }
        productImage: { type: string }
        price: { type: number }
        quantity: { type: integer }
        outOfStock: { type: boolean }
        selectedColor: { type: string, nullable: true }
    Order:
      type: object
      required: [id, userId, status, total, region, address, phone, backupPhone, items, createdAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        userName: { type: string }
        status: { $ref: "#/components/schemas/OrderStatus" }
        total: { type: number }
        deliveryFee: { type: number }
        region: { type: string }
        address: { type: string }
        phone: { type: string }
        backupPhone: { type: string }
        notes: { type: string, nullable: true }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
        createdAt: { type: string }
    UpdateOrderStatusBody:
      type: object
      required: [status]
      properties:
        status: { $ref: "#/components/schemas/OrderStatus" }
    FlagStockoutBody:
      type: object
      required: [orderItemId]
      properties:
        orderItemId: { type: integer }
...
    AdminSummary:
      type: object
      required: [totalOrders, pendingOrders, preparingOrders, totalRevenue, totalUsers, totalProducts, lowStockProducts, recentOrders]
      properties:
        totalOrders: { type: integer }
        pendingOrders: { type: integer }
        preparingOrders: { type: integer }
        totalRevenue: { type: number }
        totalUsers: { type: integer }
        totalProducts: { type: integer }
        lowStockProducts: { type: integer }
        recentOrders:
          type: array
          items: { $ref: "#/components/schemas/Order" }
    UpdateUserRoleBody:
      type: object
      required: [role]
      properties:
        role: { $ref: "#/components/schemas/UserRole" }
    CreateProductBody:
      type: object
      required: [name, price, description, imageUrl, stock, subcategoryId]
      properties:
        name: { type: string }
        price: { type: number }
        description: { type: string }
        imageUrl: { type: string }
        image2: { type: string }
        image3: { type: string }
        stock: { type: integer }
        subcategoryId: { type: integer }
        colorsEnabled: { type: boolean }
    UpdateProductBody:
      type: object
      properties:
        name: { type: string }
        price: { type: number }
        description: { type: string }
        imageUrl: { type: string }
        image2: { type: string }
        image3: { type: string }
        stock: { type: integer }
        subcategoryId: { type: integer }
        colorsEnabled: { type: boolean }
    AdminCreateCategoryBody:
      type: object
      required: [name, slug]
      properties:
        name: { type: string }
        slug: { type: string }
        nameEn: { type: string }
        icon: { type: string }
        imageUrl: { type: string }
        sortOrder: { type: integer }
    AdminUpdateCategoryBody:
      type: object
      properties:
        name: { type: string }
        slug: { type: string }
        nameEn: { type: string }
        icon: { type: string }
        imageUrl: { type: string }
        sortOrder: { type: integer }