# lib/api-zod/src/generated/types/createProductBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * متجر الأفق المتجدد API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProductBody {
  name: string;
  price: number;
  description: string;
  imageUrl: string;
  image2?: string;
  image3?: string;
  stock: number;
  subcategoryId: number;
  colorsEnabled?: boolean;
}