# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * متجر الأفق المتجدد API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  userId: number;
  userName?: string;
  status: OrderStatus;
  total: number;
  deliveryFee?: number;
  region: string;
  address: string;
  phone: string;
  backupPhone: string;
  notes?: string | null;
  items: OrderItem[];
  createdAt: string;
}