# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * متجر الأفق المتجدد API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductColor } from "./productColor";

export interface Product {
  id: number;
  name: string;
  price: number;
  description: string;
  imageUrl: string;
  image2?: string | null;
  image3?: string | null;
  colorsEnabled?: boolean;
  stock: number;
  subcategoryId: number;
  subcategorySlug: string;
  subcategoryName: string;
  categorySlug: string;
  categoryName: string;
  createdAt?: string;
  averageRating?: number | null;
  ratingCount?: number;
  colors?: ProductColor[];
}